+++ app/Models/ERP/Cashier.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\ERP;
 
 use App\Models\User;
 use App\Models\Company;
