+++ app/Models/ERP/HistoryCashier.php
@@ -1,12 +1,11 @@
 <?php
-
-namespace App\Models;
+namespace App\Models\ERP;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-/** Produto e Venda(historico) */
-class ProductSale extends Model
+/** Historico do caixa */
+class HistoryCashier extends Model
 {
     use HasFactory;
 }
