+++ app/Models/ERP/HistoryStatusSale.php
@@ -1,12 +1,11 @@
 <?php
-
-namespace App\Models;
+namespace App\Models\ERP;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-/** Movimento do caixa */
-class CashMovement extends Model
+/** Historico de status de vendas */
+class HistoryStatusSale extends Model
 {
     use HasFactory;
 }
