+++ app/Models/ERP/Product.php
@@ -1,15 +1,14 @@
 <?php
+namespace App\Models\ERP;
 
-namespace App\Models;
 
 use App\Models\User;
 use App\Models\Company;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-
-/** Contas a receber, receitas */
-class Income extends Model
+/** Produto */
+class Product extends Model
 {
     use HasFactory;
     /** Relationship */
