+++ app/Models/Financial/Expense.php
@@ -1,15 +1,14 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
 
 use App\Models\User;
 use App\Models\Company;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-
-/** Formas de pagamento */
-class PaymentMethod extends Model
+/** Despeza */
+class Expense extends Model
 {
     use HasFactory;
     /** Relationship */
