+++ app/Models/Financial/Income.php
@@ -1,6 +1,6 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
 
 use App\Models\User;
 use App\Models\Company;
@@ -8,11 +8,10 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
-/** Historico de salario */
-class SalaryHistory extends Model
+/** Contas a receber, receitas */
+class Income extends Model
 {
     use HasFactory;
-
     /** Relationship */
     public function company(): BelongsTo
     {
