+++ app/Models/Financial/Installment.php
@@ -1,6 +1,6 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
