+++ app/Models/Financial/IntervalRecurrence.php
@@ -1,12 +1,11 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-
-/** Historico de status de vendas */
-class HistoryStatusSale extends Model
+/** Recorrencia de despeza, intervalo customizado */
+class IntervalRecurrence extends Model
 {
     use HasFactory;
 }
