+++ app/Models/Financial/PaymentHistory.php
@@ -1,6 +1,6 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
 
 use App\Models\User;
 use App\Models\Company;
@@ -8,12 +8,10 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
-/** Fornecedores */
-class Supplier extends Model
+/** Historico de pagamentos, despezas, contas a pagar e a receber */
+class PaymentHistory extends Model
 {
     use HasFactory;
-
-    /** Relationship */
     public function company(): BelongsTo
     {
         return $this->belongsTo(Company::class);
