+++ app/Models/Financial/PaymentHistorySystem.php
@@ -1,24 +1,19 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
-
-use App\Models\User;
 use App\Models\Company;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
-/** Contas a pagar */
-class InvoicePay extends Model
+/** Historico de pagamento do sistema, feito pelo usuario */
+class PaymentHistorySystem extends Model
 {
     use HasFactory;
+
     /** Relationship */
     public function company(): BelongsTo
     {
         return $this->belongsTo(Company::class);
-    }
-    public function user(): BelongsTo
-    {
-        return $this->belongsTo(User::class);
     }
 }
