+++ app/Models/Financial/SalaryHistory.php
@@ -1,14 +1,14 @@
 <?php
+namespace App\Models\Financial;
 
-namespace App\Models;
-
+use App\Models\User;
 use App\Models\Company;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
-/** Historico de pagamento do sistema, feito pelo usuario */
-class PaymentHistorySystem extends Model
+/** Historico de salario */
+class SalaryHistory extends Model
 {
     use HasFactory;
 
@@ -16,5 +16,9 @@ class PaymentHistorySystem extends Model
     public function company(): BelongsTo
     {
         return $this->belongsTo(Company::class);
+    }
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
     }
 }
